Take kwargs in interest.py
Make the function `interest()` take kwargs that populate the widgets.
This allows for persisting values of a scenario in a jupyter notebook,
effectively allowing to avoid specifying desired values every time one
opens a notebook.

Also close each plot so that you can reuse the interest plots in
different scenarios.